requirements.txt
twitter-api-client>=0.10.22 
PyYAML>=6.0.1
tqdm>=4.67.1
